% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoints.R
\name{get_last_update}
\alias{get_last_update}
\alias{obter_data_atualizacao}
\title{Retrieve the data update timestamp}
\usage{
get_last_update(base_url = .obrasgovr_base_url())

obter_data_atualizacao(base_url = .obrasgovr_base_url())
}
\arguments{
\item{base_url}{HTTPS base URL. By default, uses the \code{obrasgovr.base_url}
option or the official API environment.}
}
\value{
A \code{POSIXct} value in the UTC time zone.
}
\description{
Retrieve the data update timestamp
}
\examples{
if (interactive()) {
  get_last_update()
}
}
\seealso{
Other API resources:
\code{\link[=get_commitments]{get_commitments()}},
\code{\link[=get_contracts]{get_contracts()}},
\code{\link[=get_feasibility_studies]{get_feasibility_studies()}},
\code{\link[=get_geometries]{get_geometries()}},
\code{\link[=get_physical_execution]{get_physical_execution()}},
\code{\link[=get_projects]{get_projects()}},
\code{\link[=get_status_history]{get_status_history()}}
}
\concept{API resources}
