% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{obrasgov_recursos}
\alias{obrasgov_recursos}
\title{Recursos disponiveis na API ObrasGov}
\usage{
obrasgov_recursos()
}
\value{
Um tibble com o nome do recurso, a funcao correspondente e o
endpoint da API.
}
\description{
Recursos disponiveis na API ObrasGov
}
\examples{
obrasgov_recursos()
}
